feat : Coordinator 기본 설정
