+++ FanFollow/FanFollow/Source/Coordinator/SettingCoodinator.swift
@@ -1,0 +1,26 @@
+//
+//  SettingCoodinator.swift
+//  FanFollow
+//
+//  Copyright (c) 2023 Minii All rights reserved.
+
+import UIKit
+
+class SettingCoordinator: Coordinator {
+    weak var parentCoordinator: MainTabBarCoordinator?
+    var childCoordinators: [Coordinator] = []
+    var navigationController: UINavigationController
+    
+    init(navigationController: UINavigationController) {
+        self.navigationController = navigationController
+    }
+    
+    func start() {
+        let controller = SettingTabBarController()
+        controller.coordinator = self
+        
+        navigationController.pushViewController(controller, animated: true)
+    }
+    
+    func presentSettingViewController() { }
+}
