--- FanFollow/FanFollow/Source/Domain/UseCase/ExampleUseCase.swift
@@ -1,21 +1,0 @@
-//
-//  ExampleUseCase.swift
-//  FanFollow
-//
-//  Copyright (c) 2023 Minii All rights reserved.
-
-/// 해당 파일은 예시 파일입니다. 삭제 하셔도 됩니다.
-
-import RxSwift
-
-protocol ExampleUseCase: AnyObject {
-//    func fetchUserInformationData() -> Observable<UserInformation>
-}
-
-final class DefaultExampleUseCase: ExampleUseCase {
-//    private let postRepository: PostRepository
-//
-//    init(postRepository: PostRepository) {
-//        self.postRepository = postRepository
-//    }
-}
