+++ FanFollow/FanFollow/Source/Presenter/Comman/UI/Controller/TopTabBarController.swift
@@ -21,6 +21,12 @@ class TopTabBarController<T: TabItem>: UITabBarController {
         bind()
     }
     
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        
+        navigationController?.setNavigationBarHidden(true, animated: false)
+    }
+    
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
