--- FanFollow/FanFollow/Source/Presenter/MainScene/View/TabBarController.swift
@@ -1,24 +1,0 @@
-//
-//  TabBarController.swift
-//  FanFollow
-//
-//  Copyright (c) 2023 Minii All rights reserved.
-
-import UIKit
-
-//final class TabBarController: UITabBarController {
-//    private let routers = Router.allCases
-//
-//    weak var coordinator: TabBarCoordinator?
-//
-//    override func viewDidLoad() {
-//        super.viewDidLoad()
-//        view.backgroundColor = .systemBackground
-//
-//
-//    }
-//}
-//
-//private extension TabBarController {
-//
-//}
